Project version 1.3 added controller details expenses family with test (falling two tests in ControllerTests)

+++ src/main/java/ru/top/java212/family/AllExpensesFamily.java
@@ -6,5 +6,5 @@ import java.util.Map;
 public interface AllExpensesFamily {
     int calculationExpensesFamily(LocalDate initalDate, LocalDate endData);
 
-    Map<String, Integer> getExpensesByCategory(LocalDate initalDate, LocalDate endData);
+    Map<String, Long> getExpensesByCategory(LocalDate initalDate, LocalDate endData);
 }
